adds new file called StarRatingStackView into Views folder
