+++ BetterReads/BetterReads/View Controllers/BookDetailViewController.swift
@@ -168,6 +168,8 @@ class BookDetailViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        title = ""
+        print("intrinsicContentSize = \(self.view.intrinsicContentSize)")
         setupSubviews()
     }
     
@@ -183,6 +185,12 @@ class BookDetailViewController: UIViewController {
         navigationController?.navigationBar.barStyle = .default
     }
     
+    // FIXME: button might be weird because contentSize is -1, -1
+//    override var intrinsicContentSize: CGSize {
+//      //let width = componentsWidth + componentsSpacing
+//        return CGSize(width: view.bounds.width, height: view.bounds.height)
+//    }
+    
     // FIXME: THIS NEEDS TO RUN BEFORE THIS VIEW LOADS. USE NOTIFICATION TO PASS TOP BAR HEIGTH BEFOREHAND
     private func statusAndNavigationBarHeight() -> CGFloat {
         let navHeight = self.navigationController?.navigationBar.frame.height ?? 0.0
@@ -195,7 +203,6 @@ class BookDetailViewController: UIViewController {
     private func setupSubviews() {
         // Scroll View
         // add the scroll view to self.view
-        title = ""
         self.view.addSubview(scrollView)
         // constrain the scroll view to 8-pts on each side
         scrollView.topAnchor.constraint(equalTo: view.topAnchor).isActive = true
@@ -326,6 +333,9 @@ class BookDetailViewController: UIViewController {
         dummyView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -16).isActive = true
         dummyView.widthAnchor.constraint(equalToConstant: 60).isActive = true
         dummyView.heightAnchor.constraint(equalToConstant: 30).isActive = true
+        
+        print("intrinsicContentSize at end = \(self.view.intrinsicContentSize)")
+
     }
     
     
@@ -339,71 +349,4 @@ class BookDetailViewController: UIViewController {
      }
      */
     
-}
-
-class StarRatingStackView: UIStackView {
-    
-    /// Holds 5 star images insied
-    var starsView: UIView!
-    
-    /// Array of star images
-    var starsArray = [UIImageView]()
-    
-    // FIXME: give this a didSet later that calls updateStarRating
-    /// How many stars should be filled in
-    var ratingValue: Double? {
-        didSet {
-            updateStarRating()
-        }
-    }
-    var starSize: Double = 20.0
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        print("init with frame")
-        setupSubviews()
-    }
-    
-    required init(coder: NSCoder) {
-        super.init(coder: coder)
-        print("init with coder")
-        setupSubviews()
-    }
-    
-    /// Sets up stackView and stars
-    private func setupSubviews() {
-        
-        distribution = .fillEqually
-        //let starSize = Int(20) // FIXME: should be based on view size?
-        for i in 1...5 {
-            let star = UIImageView()
-            star.translatesAutoresizingMaskIntoConstraints = false
-            addArrangedSubview(star)
-            starsArray.append(star)
-            star.tag = i
-            star.contentMode = .scaleToFill
-            star.frame = CGRect(x: 0,
-                                y: 0,
-                                width: starSize,
-                                height: starSize)
-            star.image = UIImage(named: "Stars_Chunky-AltoGray")
-        }
-        updateStarRating()
-    }
-    
-    private func updateStarRating() {
-        
-        let value = ratingValue ?? 0
-        var chunk = value
-        for star in starsArray {
-            if chunk >= 0.66 && chunk <= 5.0 {
-                star.image = UIImage(named: "Stars_Chunky-DoveGray")
-            } else if chunk >= 0.33 && chunk < 0.66 {
-                star.image = UIImage(named: "Stars_Chunky-AltoGray-LeftHalf")
-            } else {
-                star.image = UIImage(named: "Stars_Chunky-AltoGray")
-            }
-            chunk = value - Double(star.tag)
-        }
-    }
 }
